+++ contracts/analytics/TrancheReserveData.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.8.0;
 import {ILendingPool} from "../interfaces/ILendingPool.sol";
 
 contract TrancheReserveData {
-    constructor(address pool, uint16 trancheId) {
+    constructor(address pool, uint64 trancheId) {
         // ILendingPool(pool).getReserveData(asset, trancheId)
         address[] memory list = ILendingPool(pool).getReservesList(trancheId);
         bytes memory returnData = abi.encode(list);
